% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-av.r
\name{search_avs}
\alias{search_avs}
\title{Search anti-virus indicators}
\usage{
search_avs(avs, pause = 10)
}
\arguments{
\item{avs}{character vector of AV indicators to search for}

\item{pause}{ThreatCrowd requests a 10s pause between requests. A higher
or lower value can be specified but going lower may get your IP banned.}
}
\description{
Search anti-virus indicators
}
\examples{
search_avs("plugx")
}

