% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-domain.r
\name{search_domains}
\alias{search_domains}
\title{Search domain indicators}
\usage{
search_domains(domains, pause = 10)
}
\arguments{
\item{domains}{character vector of domains to search for}

\item{pause}{ThreatCrowd requests a 10s pause between requests. A higher
or lower value can be specified but going lower may get your IP banned.}
}
\description{
Search domain indicators
}
\examples{
search_domains("aoldaily.com")
}

