% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-email.r
\name{search_email}
\alias{search_email}
\title{Search e-mail indicators}
\usage{
search_email(emails, pause = 10)
}
\arguments{
\item{emails}{character vector of e-mails to search for}

\item{pause}{ThreatCrowd requests a 10s pause between requests. A higher
or lower value can be specified but going lower may get your IP banned.}
}
\description{
Search e-mail indicators
}
\examples{
search_email("william19770319@yahoo.com")
}

