% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-ip.r
\name{search_ips}
\alias{search_ips}
\title{Search IP indicators}
\usage{
search_ips(ips, pause = 10)
}
\arguments{
\item{ips}{character vector of IP addresses to search for}

\item{pause}{ThreatCrowd requests a 10s pause between requests. A higher
or lower value can be specified but going lower may get your IP banned.}
}
\description{
Search IP indicators
}
\examples{
search_ips("188.40.75.132")
}

